03.04.2021 - Admin POST(linkTo) - RD

+++ src/main/java/com/qvik/events/infra/response/dto/LinkToDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class LinkToDTO {
 
-	private Integer sourceId;
-	private Integer destinationId;
+	private Long sourceId;
+	private Long destinationId;
 }
